fix(web): addressing requested changes
