+++ Web/starter_project_vanguard/data/about/a2sv-sessions.ts
@@ -1,36 +1,9 @@
-import hakimHub from '../../public/aboutus/hakimhub.svg'
-import howToApproach from '../../public/aboutus/how_to_approach.svg'
-import problemSolvingSession from '../../public/aboutus/problem_solving_session.svg'
-import qAndA from '../../public/aboutus/q&a.svg'
-import technicalTraining from '../../public/aboutus/technical_training.svg'
-import trackSys from '../../public/aboutus/tracksys.svg'
-import weeklyContests from '../../public/aboutus/weekly_contests.svg'
-import weeklyOneOnOnes from '../../public/aboutus/weekly_one_on_ones.svg'
-
-export const socialProjects = [
-  {
-    image: hakimHub,
-    title: 'Hakim Hub',
-    content:
-      'HakimHub is a platform that provides information about healthcare\
-            facilities and healthcare professionals in Ethiopia. Hakimhub makes\
-            information about hospitals, medical laboratories, and doctors\
-            conveniently accessible to its users.',
-    leftAligned: false,
-    isImageLeft: true,
-  },
-  {
-    image: trackSys,
-    title: 'Track Sym',
-    content:
-      'TrackSym is a non-commercial app that uses crowd-sourcing to collect\
-              and visualize the density of the relevant Covid-19 symptoms. Symptom\
-              data, aggregated by places, can help people avoid visiting areas\
-              that are heavily populated by symptomatic people.',
-    leftAligned: true,
-    isImageLeft: false,
-  },
-]
+import howToApproach from '../../public/images/aboutus/sessions/how-to-approach.svg'
+import problemSolvingSession from '../../public/images/aboutus/sessions/problem-solving-session.svg'
+import qAndA from '../../public/images/aboutus/sessions/q&a.svg'
+import technicalTraining from '../../public/images/aboutus/sessions/technical-training.svg'
+import weeklyContests from '../../public/images/aboutus/sessions/weekly-contests.svg'
+import weeklyOneOnOnes from '../../public/images/aboutus/sessions/weekly-one-on-ones.svg'
 
 export const a2svSessions = [
   {
